feat- divide the Code of PreviewComparison.tsx into three components
